+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - **[NEW]** Brand new `Itinerary` component (under `newItinerary` folder)
 [...]
 
+- **[FIX]** Remove background from `SEA Pages`
+
 # v47.5.0 (12/02/2021)
 
 - **[FIX]** Fix `CheckCircleIcon` display when it has a title