+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 [...]
 
 - **[FIX]** Fix `Success Screen` gutter for large screens
+- **[FIX]** Remove background from `SEA Pages`
+
 
 # v47.5.0 (12/02/2021)