+++ CHANGELOG.md
@@ -5,7 +5,9 @@
 - **[NEW]** Add `BrandedVoice` component
 [...]
 
+- **[FIX]** Fix `Success Screen` gutter for large screens
 - **[FIX]** Remove background from `SEA Pages`
+
 
 # v47.5.0 (12/02/2021)